warnings: fix signed/unsigned comparisons

+++ hlslang/GLSLCodeGen/hlslLinker.cpp
@@ -1391,7 +1391,7 @@ void HlslLinker::appendDuplicatedInSemantics(GlslSymbolOrStructMemberBase* sym, 
 void HlslLinker::markDuplicatedInSemantics(GlslFunction* func)
 {
 	int pCount = func->getParameterCount();
-	for (int asi=0; asi < sizeof(kAttributeSemantic)/sizeof(kAttributeSemantic[0]); ++asi)
+	for (size_t asi=0; asi < sizeof(kAttributeSemantic)/sizeof(kAttributeSemantic[0]); ++asi)
 	{
 		std::vector<GlslSymbolOrStructMemberBase*> symsUsingSem;
 		AttrSemanticMapping* sem = &kAttributeSemantic[asi];
@@ -1402,7 +1402,7 @@ void HlslLinker::markDuplicatedInSemantics(GlslFunction* func)
 		}
 		if (symsUsingSem.size() > 1)
 		{
-			int index_of_largest = -1;
+			size_t index_of_largest = -1;
 			int largest_array_size = 0;
 			GlslSymbolOrStructMemberBase* sym_of_largest = 0;
 			for (unsigned int ii=0; ii < symsUsingSem.size(); ii++)
@@ -1414,7 +1414,7 @@ void HlslLinker::markDuplicatedInSemantics(GlslFunction* func)
 					largest_array_size = getElements(symsUsingSem[ii]->type);
 				}
 			}
-			for (unsigned int ii=0; ii < symsUsingSem.size(); ii++)
+			for (size_t ii=0; ii < symsUsingSem.size(); ii++)
 			{
 				if (ii != index_of_largest)
 				{
